refactor GNN to allow any output (inputs are always the same)
one step toward #731
